Move NoteInteractors.kt
