+++ .idea/dictionaries/luthfi.xml
@@ -1,0 +1,7 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="luthfi">
+    <words>
+      <w>interactors</w>
+    </words>
+  </dictionary>
+</component>
