+++ app/src/main/java/com/laam/laamnotes/framework/interactors/NoteInteractors.kt
@@ -1,11 +1,11 @@
-package com.laam.laamnotes.framework
+package com.laam.laamnotes.framework.interactors
 
 import com.laam.core.usecase.AddNote
 import com.laam.core.usecase.GetAllNotes
 import com.laam.core.usecase.GetNote
 import com.laam.core.usecase.RemoveNote
 
-data class UseCases(
+data class NoteInteractors(
     val addNote: AddNote,
     val getAllNotes: GetAllNotes,
     val getNote: GetNote,
